+++ GameCore/OpenGlHelper/ObjectVectors.cs
@@ -1,0 +1,23 @@
+﻿#region
+
+using OpenGL;
+
+#endregion
+
+namespace GameCore.OpenGlHelper
+{
+    /// <summary>
+    /// This holds the vectors that create object primitives.
+    /// </summary>
+    public class ObjectVectors
+    {
+        public Vector3[] Vertex;
+
+        /// <summary>
+        ///     the indexed of the vertexes that create triangles.
+        /// </summary>
+        public int[] ElementData;
+
+        public Vector2[] Uvs;
+    }
+}