+++ ChangeLog.md
@@ -1,4 +1,6 @@
 
+2015.02.02	IMPROVEMENTS	The console window is now placed on the secondary screen.
+2015.02.02	IMPROVEMENTS	Implemented an ObjectPrimitives class that helps create cubes and squares.
 2015.02.01	IMPROVEMENTS	Implemented a semi transparent bitmap for the HUD Panel.
 2015.02.01	IMPROVEMENTS	Added keyBindings class. The keyboard controlles can now easily be remapped.
 2015.01.31	IMPROVEMENTS	All console messages are now broadcasted via the GameCore.TheGameCore.RaiseMessage() method.